requirements.txt
openai
streamlit
streamlit-lottie
python-dotenv
PyPDF2
rank_bm25
nltk
python-docx
pdfplumber
pytesseract
pymupdf
sentence-transformers
Pillow
python-dotenv
python-pptx
